Exclude syslog from windows targets.
